Docker-Django/django_project/requirements.txt
Django==4.2
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.15.1
pillow==10.2.0